+++ CHANGELOG.md
@@ -135,6 +135,7 @@ Modifications can only be done via builders.
 - Make removeSupplier private. Only dispose and erase can be used.
 - Make removeCustomer private. Only dispose and erase can be used.
 - Make erase private. Publicly nodes can only be disposed. Erasal happens when the last customer is removed from a node.
+- Mute suppliers in the blue print when a node is disposed
 
 ### Fixed